AP_OpticalFlow: move OpticalFlow.h to AP_OpticalFlow.h
